Move context to protocol module
