+++ src/types/stardust/tangle/block/output/mod.rs
@@ -36,9 +36,10 @@ pub use self::{
     },
 };
 use super::Address;
-use crate::types::{
-    context::{TryFromWithContext, TryIntoWithContext},
-    stardust::{tangle::block::payload::transaction::TransactionId, ProtocolParameters},
+use crate::types::stardust::{
+    protocol::{TryFromWithContext, TryIntoWithContext},
+    tangle::block::payload::transaction::TransactionId,
+    ProtocolParameters,
 };
 
 /// The amount of tokens associated with an output.
