+++ src/types/stardust/tangle/block/payload/milestone/mod.rs
@@ -12,12 +12,14 @@ use serde::{Deserialize, Serialize};
 
 pub use self::milestone_id::MilestoneId;
 use crate::types::{
-    context::{TryFromWithContext, TryIntoWithContext},
-    stardust::tangle::{
-        block::{payload::treasury_transaction::TreasuryTransactionPayload, Address, BlockId, Signature},
-        milestone::MilestoneIndex,
-    },
     serde::bytify,
+    stardust::{
+        protocol::{TryFromWithContext, TryIntoWithContext},
+        tangle::{
+            block::{payload::treasury_transaction::TreasuryTransactionPayload, Address, BlockId, Signature},
+            milestone::MilestoneIndex,
+        },
+    },
 };
 
 /// Represents a milestone payload.
